Move endpoints to 'api' folder
