+++ flaskr/api/course/delete_course.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 from flask import jsonify
-from flaskr.course import bp
+from flaskr.api.course import bp
 from flaskr.db import db
 from flaskr.utils import exists_course, login_required
