+++ flaskr/api/course/get_course_by_id.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 from flask import jsonify
-from flaskr.course import bp
+from flaskr.api.course import bp
 from flaskr.db import CourseSchema
 from flaskr.utils import exists_course, login_required
