+++ flaskr/api/course/get_courses.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 from flask import jsonify
-from flaskr.course import bp
+from flaskr.api.course import bp
 from flaskr.db import Course, CourseSchema
 from flaskr.utils import login_required
