+++ flaskr/api/user/get_user.py
@@ -2,7 +2,7 @@
 
 from flask import jsonify
 from flaskr.db import UserSchema
-from flaskr.user import bp
+from flaskr.api.user import bp
 from flaskr.utils import login_required
